implement category delete service

+++ category-microservice/src/main/java/dongcom/service/impl/CategoryServiceImpl.java
@@ -43,8 +43,11 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategoryById(Long id) throws Exception {
+    public void deleteCategoryById(Long id, Long studyId) throws Exception {
         Category category = getCategoryById(id);
+        if (category.getStudyId().equals(studyId)) {
+            throw new Exception("You don't have permission to delete this category");
+        }
         categoryRepository.deleteById(id);
     }
 }
